# book-social-network/book-netwok-api/src/main/java/com/ali/book/auth/AuthenticationController.java
package com.ali.book.auth;


import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
//import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.mail.MessagingException;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;


@RestController
@RequestMapping("auth")
@RequiredArgsConstructor
@Tag(name = "Authentication")
public class AuthenticationController {

	private final AuthenticationService service;
	
	@PostMapping("/register")
	@ResponseStatus(HttpStatus.ACCEPTED)
	public ResponseEntity<?> register(@RequestBody @Valid RegistrationRequest request) throws MessagingException{
		service.register(request);
		return ResponseEntity.accepted().build();
	}
	
	@PostMapping("/authenticate")
	public ResponseEntity<AuthenticationResponse> authenticate(@RequestBody @Valid AuthenticationRequest request)
	{
		return ResponseEntity.ok(service.authenticate(request));
	}	
	@GetMapping("/activate-account")
	public void confirm(@RequestParam String token) throws MessagingException {
		service.activateAccount(token);
	}
	  @PostMapping("/forgot-password")
	    public ResponseEntity<?> forgotPassword(@RequestParam String email) throws MessagingException {
	        service.sendPasswordResetEmail(email);
	        return ResponseEntity.ok().build();
	    }

	    @PostMapping("/reset-password")
	    public ResponseEntity<?> resetPassword(@RequestParam String token, @RequestParam String newPassword) {
	        service.resetPassword(token, newPassword);
	        return ResponseEntity.ok().build();
	    }
}