# book-social-network/book-netwok-api/src/main/java/com/ali/book/auth/AuthenticationRequest.java
package com.ali.book.auth;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationRequest {

	@Email(message = "Email is not formated")
	@NotEmpty(message = "email is mandatory")
	@NotBlank(message = "email is mandatory")
	private String email;
	@Size(min = 8, message = "password should be 8 characters long minimum")
	@NotEmpty(message = "password is mandatory")
	@NotBlank(message = "password is mandatory")
	private String password;
}