# book-social-network/book-netwok-api/src/main/java/com/ali/book/auth/RegistrationRequest.java
package com.ali.book.auth;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class RegistrationRequest {

	@NotEmpty(message = "First name is mandatory")
	@NotBlank(message = "First name is mandatory")
	private String firstname;
	@NotEmpty(message = "Last name is mandatory")
	@NotBlank(message = "Last name is mandatory")
	private String lastname;
	@Email(message = "Email is not formated")
	@NotEmpty(message = "email is mandatory")
	@NotBlank(message = "email is mandatory")
	private String email;
	@Size(min = 8, message = "password should be 8 characters long minimum")
	@NotEmpty(message = "password is mandatory")
	@NotBlank(message = "password is mandatory")
	private String password;
}